requirements.txt
pytest
allure-pytest
requests
jsonschema
curlify
python-dotenv